Add title screen. Add story picker. Refactored into scenes.
