+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "type-run",
+  "name": "eat-my-dust",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
