+++ public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>TypeRun</title>
+  <title>Eat My Dust!</title>
 </head>
 <body>
   <canvas width="960" height="540"></canvas>
@@ -28,5 +28,33 @@
   </div>
 
   <div id="div-label"></div>
+
+  <div id="div-title">
+    <div class="title">Eat My Dust!</div>
+  </div>
+
+  <div id="div-credits">
+    <div class="developers">
+      <div>Developed by</div>
+      <ul>
+        <li>Junsik Shim</li>
+        <li>Juhyun Choi</li>
+      </ul>
+    </div>
+    <div class="artists">
+      <div>Artworks by</div>
+      <ul>
+        <li>wabie</li>
+      </ul>
+    </div>
+  </div>
+
+  <div id="div-message" class="blink">Press Space to start!</div>
+
+  <div id="div-story-picker">
+    <div>Select a story</div>
+    <ul></ul>
+    <div>Up/Down arrows to choose, Enter to select</div>
+  </div>
 </body>
 </html>
